Add files via upload
Hi, it would not let me commit or push my final edit after finishing the extra credit function. I uploaded this as my final version of the python file. I hope this works. If not, I understand if I am unable to receive the extra credit but I hope this is something that can be looked passed. Thank you!! ♡